Completed twitch stream v2
The script will now show a random streamer when no-one is streaming
(could be nice for channel promotion :+1: ). When only one user is live,
that user will be shown on the player. If multiple users are live, a
random pick will be made for who is shown on the player. Users can click
the boxes below the player at any time to load a different streamer.
Some errors are shown in the console, but I *suspect* those come from
Twitch itself, not from the actual javascript.